Split chapter subcomponent tests into focused files
